requirements.txt
asgiref==3.7.2
Django==4.2.3
graphql-core==3.2.3
psycopg2-binary==2.9.7
python-dateutil==2.8.2
six==1.16.0
sqlparse==0.4.4
strawberry-graphql==0.204.0
strawberry-graphql-django==0.16.0
typing_extensions==4.7.1
